+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -2,6 +2,7 @@ Version X.X.X
 =============
 
  * 2016-07-08 Security fix against DoS in Commons FileUpload (CVE-2016-3092) (suer)
+ * 2016-07-04 Fixed directory listing in AssetsController (mallowlabs)
 
 Version 5.6.0
 =============